pyproject.toml
[tool.poetry]
name = "tz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
